still not properly setting the Listener to the Fragment
