+++ app/src/main/java/com/jbrys/android/bloquery/ui/activity/BloQueryActivity.java
@@ -4,9 +4,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.app.AppCompatDialogFragment;
 import android.support.v7.widget.Toolbar;
-import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
@@ -15,7 +13,6 @@ import android.widget.TextView;
 import com.jbrys.android.bloquery.BloQueryApplication;
 import com.jbrys.android.bloquery.R;
 import com.jbrys.android.bloquery.api.model.Question;
-import com.jbrys.android.bloquery.ui.dialog.AnswerQuestionDialog;
 import com.jbrys.android.bloquery.ui.fragment.QuestionDetailFragment;
 import com.jbrys.android.bloquery.ui.fragment.QuestionsFragment;
 import com.parse.ParseUser;
@@ -25,13 +22,11 @@ import com.parse.ParseUser;
  */
 public class BloQueryActivity extends AppCompatActivity implements
         Button.OnClickListener,
-        QuestionsFragment.Listener,
-        AnswerQuestionDialog.AnswerDialogListener{
+        QuestionsFragment.Listener{
     private final String TAG = getClass().getSimpleName();
 
     private ParseUser currentUser;
 
-    private Menu mMenu;
     private Toolbar mToolbar;
     private TextView userNameTextView;
     private Button logout;
@@ -77,22 +72,6 @@ public class BloQueryActivity extends AppCompatActivity implements
         }
     }
 
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.bloquery, menu);
-        this.mMenu = menu;
-        return super.onCreateOptionsMenu(menu);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-
-        if (item.getItemId() == R.id.menu_answer){
-            showAnswerQuestionDialog();
-        }
-
-        return super.onOptionsItemSelected(item);
-    }
 
     @Override
     public void onBackPressed() {
@@ -126,9 +105,6 @@ public class BloQueryActivity extends AppCompatActivity implements
 
     @Override
     public void onItemAnswersClicked(QuestionsFragment questionsFragment, Question question) {
-        answerItem = mMenu.findItem(R.id.menu_answer);
-        answerItem.setVisible(true);
-        answerItem.setEnabled(true);
 
         getFragmentManager().beginTransaction()
                 .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE)
@@ -138,21 +114,7 @@ public class BloQueryActivity extends AppCompatActivity implements
 
     }
 
-    public void showAnswerQuestionDialog() {
-        AnswerQuestionDialog dialog = new AnswerQuestionDialog();
-        dialog.show(getSupportFragmentManager(), "answer");
-    }
 
 
-    @Override
-    public void onDialogPositiveClick(AppCompatDialogFragment dialog, Bundle args) {
-        String answerText;
 
-        answerText = args.getString("questionText");
-    }
-
-    @Override
-    public void onDialogNegativeClick(AppCompatDialogFragment dialog) {
-
-    }
 }
