+++ app/src/main/res/menu/question_detail.xml
@@ -6,7 +6,7 @@
         android:id="@+id/menu_answer"
         android:icon="@drawable/ic_lightbulb_outline_white_36dp"
         android:title="Answer Question"
-        android:visible="false"
+        android:visible="true"
         app:showAsAction="always" />
 
 </menu>
